fix(vue): restore raw input resolution
Restore the Vue resolver removed from the v2 client during the core rewrite. Raw entries resolve Vue values during rendering while automatic dependency tracking remains in useHead().
